+++ CHANGELOG.md
@@ -57,6 +57,9 @@ Increment the:
 * [CODE HEALTH] Fix clang-tidy unnecessary value param warnings
   [#3931](https://github.com/open-telemetry/opentelemetry-cpp/pull/3931)
 
+* [EXAMPLE] Add manual asynchronous context propagation example
+  [#3935](https://github.com/open-telemetry/opentelemetry-cpp/pull/3935)
+
 Important changes:
 
 * [BUILD] Revisit EventLogger deprecation